// web/src/api/models/settings.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ATRAG API
 * ATRAG API Documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface Settings
 */
export interface Settings {
    /**
     * Whether to use MinerU
     * @type {boolean}
     * @memberof Settings
     */
    'use_mineru'?: boolean;
    /**
     * API token for MinerU
     * @type {string}
     * @memberof Settings
     */
    'mineru_api_token'?: string;
    /**
     * Whether to use DocRay
     * @type {boolean}
     * @memberof Settings
     */
    'use_doc_ray'?: boolean;
    /**
     * Whether to use MarkItDown
     * @type {boolean}
     * @memberof Settings
     */
    'use_markitdown'?: boolean;
}